rename spawn_task_many.hpp to spawn_many.hpp
